Loads of changes here...maybe TOO many! On the VAE front, we now have both the regular and "spatially enhanced" versions running. The MPS usage is still quite low, this might be because:
- Batch size = 128
- Input dimension = 375 floats per sample
- Total elements in batch = 128 × 375 = 48,000 elements

Which is about 187 KB. This might be why we see such low usage. The timing examples are included below as the "before". By my math we would need to increas batch size considerably just to get to 1GB (something approaching 4 orders of magnitude). I'll experiment with that on the next runs...after we reach a save point.